+++ changelog.md
@@ -1,2 +1,3 @@
 - Fixed 1.21.1 crash when placing blocks with containers
 - Fixed desk displaying the wrong tooltip (#112)
+- Added colored items to dye tags